requirements.txt
groq
langchain
langchain-groq
langchain-community
python-dotenv